+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Add `find_all` static method to `TransactionLineItem` class
 * Add support for `profile_id` in Transaction#create options for VenmoAccounts
 
-## 2.82.0
+# 2.82.0
 * Add support for tagged evidence in DisputeGateway#add_text_evidence (Beta release)
 * Update https certificate bundle